[nemo-qml-plugin-notification] Separating notification data from Notification class
This commit introduces a NotificationData class, that is used to store
attributes for Notification class and carry the data of a Notification
through DBus.

This modification fixes the inability for the compiler to compile the module,
especially in debug mode, as Notification inherits from QObject, and QObject don't
have an assignment operator, while this operator is required by QList to work.