[nemo-qml-plugin-notifications] Documentation improvements. Contributes to MER#1157
Add documentation for the C++ interface class, and bring the QML type
documentation up to date.

Note that qdoc does not correctly handle a module having a QML type
and a C++ class with matching names.  To work around this problem, the
C++ is documented inside a fictitious 'libnemonotifications' namespace
which avoids the name conflict.